Split store into multiple files
